refactor site selector
