+++ src/ros/nova-gui/nova-gui/src/components/nir-probe/NIRProbeOutputWidget/NIRProbeWidget.tsx
@@ -1,4 +1,5 @@
 import React, {useCallback, useState} from "react";
+import SiteSelectWidget from "../../SiteSelectWidget/SiteSelectWidget.tsx";
 import NIRProbeOutputSaveWidget from "./NIRProbeOutputSaveWidget.tsx";
 import NIRProbeLEDWidget from "../NIRProbeLEDWidget/NIRProbeLEDWidget.tsx";
 import NIRProbeFileTableWidget from "./NIRProbeFileTableWidget.tsx";
@@ -6,11 +7,10 @@ import NIRCalibrationCurveWidget, {EMPTY_CALIBRATION_DATA, NIRCalibrationData} f
 import {useLocalStorage} from "../../../hooks/useLocalStorage.ts";
 import TOFHeight from "../../AnalysisPlatformHeight/AnalysisPlatformHeight.tsx";
 import {DEFAULT_WATER_CALIBRATION} from "./DefaultWaterCalibration.ts";
-import {useGenericStore} from "../../../hooks/useGenericStore.ts";
+import {useGenericStore} from "../../../redux/actions/useGenericStore.ts";
 import {Site} from "../../../redux/models/genericStores/CurrentSiteStore.ts";
 import {SiteData, SiteDataState} from "../../../redux/models/genericStores/SiteDataState.ts";
 import {SpaceResourcesSiteType} from "../SpaceResourcesSiteType.tsx";
-import SiteTypeSelectWidget from "../../SiteSelectWidget/SiteTypeSelectWidget.tsx";
 
 interface INIRProbeWidgetProps {
 }
@@ -60,7 +60,7 @@ const NIRProbeWidget: React.FC<INIRProbeWidgetProps> = () => {
       </div>
       <div className="flex flex-col gap-3 col-span-3">
         <TOFHeight/>
-        <SiteTypeSelectWidget/>
+        <SiteSelectWidget/>
         <NIRProbeFileTableWidget file={file} setFile={setFileForCurrentSite} showAdvanced={showAdvanced} absorbance={absorbance} calibrationFunction={calibrationFunction}></NIRProbeFileTableWidget>
       </div>
     </div>
