+++ src/ros/nova-gui/nova-gui/src/redux/models/genericStores/SiteDataState.ts
@@ -1,12 +1,10 @@
 import {Site} from "./CurrentSiteStore.ts";
 import {ISpaceResourcesEntry, SpaceResourcesSiteType} from "../../../components/nir-probe/SpaceResourcesSiteType.tsx";
-import {ThresholdingFileEntry} from "../../../components/MicroscopeThresholdWidget/MicroscopeThresholdWidget.tsx";
 
 // Data that is associated with a site can be stored here.
 export interface SiteData {
   siteType: SpaceResourcesSiteType,
   spaceResourcesEntries: ISpaceResourcesEntry[],
-  thresholdingEntries: ThresholdingFileEntry[]
 }
 
 export interface SiteDataState {
@@ -20,10 +18,9 @@ export interface SiteDataState {
 const EMPTY_SITE_DATA: SiteData = {
   siteType: SpaceResourcesSiteType.WATER,
   spaceResourcesEntries: [],
-  thresholdingEntries: []
 };
 
-export const initialSiteDataState: SiteDataState = {
+export const initialNIRProbeFilesState: SiteDataState = {
   [Site.SITE_1]: EMPTY_SITE_DATA,
   [Site.SITE_2]: EMPTY_SITE_DATA,
   [Site.SITE_3]: EMPTY_SITE_DATA,
