+++ src/ros/nova-gui/nova-gui/src/views/test/TestStateView/SiteSelectorTest.tsx
@@ -1,7 +1,9 @@
 import {Card, CardBody} from "@nextui-org/react";
-import {useGenericStore} from "../../../hooks/useGenericStore.ts";
+import {useGenericStore} from "../../../redux/actions/useGenericStore.ts";
 import {Site} from "../../../redux/models/genericStores/CurrentSiteStore.ts";
 import SiteTypeSelectWidget from "../../../components/SiteSelectWidget/SiteTypeSelectWidget.tsx";
+import {siteFilenames} from "../../../components/SiteSelectWidget/OldSiteSelectWidget.tsx";
+
 
 export default function SiteSelectorTest() {
 
@@ -13,7 +15,7 @@ export default function SiteSelectorTest() {
         <SiteTypeSelectWidget/>
       </CardBody>
       <CardBody className="text-center">
-        {`site ${currentSite.valueOf() + 1}`}
+        {siteFilenames[currentSite]}
       </CardBody>
     </Card>
   )
